now is not possible to delete or post comments if the user has not the permissions

+++ app/src/api/controllers/APICommentController.php
@@ -35,30 +35,36 @@ class APICommentController extends APIController
 
     public function post($params = null)
     {
-        $data = $this->getData();
-        $comment = $this->model->get($data['subject_id'], $data['user_id']);
-        if (!empty($comment)) {
-            $this->view->response("El ususario ya ha comentado esta materia", 409);
-        } else {
-            $commentId = $this->model->create($data);
-            if (!empty($commentId))
-                $this->view->response($commentId, 201);
-            else {
-                $this->view->response("No se pudo crear el comentario", 500);
+        if (AuthHelper::checkLoggedIn()) {
+            $data = $this->getData();
+            $comment = $this->model->get($data['subject_id'], $data['user_id']);
+            if (!empty($comment)) {
+                $this->view->response("El ususario ya ha comentado esta materia", 409);
+            } else {
+                $commentId = $this->model->create($data);
+                if (!empty($commentId))
+                    $this->view->response($commentId, 201);
+                else {
+                    $this->view->response("No se pudo crear el comentario", 500);
+                }
             }
-        }
+        } else
+            $this->view->response("El usuario no esta logeado", 401);
     }
 
     public function delete($params = null)
     {
-        $commentId = $params[':ID'];
-        if ($this->model->get($commentId)) {
-            if ($this->model->delete($commentId))
-                $this->view->response("Comentario eliminado", 200);
-            else
-                $this->view->response("No se pudo eliminar el comentario", 500);
-        } else {
-            $this->view->response("No se encontró el comentario", 404);
-        }
+        if (AuthHelper::checkAdmin()) {
+            $commentId = $params[':ID'];
+            if ($this->model->get($commentId)) {
+                if ($this->model->delete($commentId))
+                    $this->view->response("Comentario eliminado", 200);
+                else
+                    $this->view->response("No se pudo eliminar el comentario", 500);
+            } else {
+                $this->view->response("No se encontró el comentario", 404);
+            }
+        } else
+            $this->view->response("El usuario no esta logeado o no tiene permisos para eliminar comentarios", 401);
     }
 }
